🧭 D8EM8Y tests: add route registry
