+++ .agentplane/tasks/202605010645-D8EM8Y/pr/github-body.md
@@ -1,0 +1,33 @@
+## Summary
+
+AP-07: Introduce unified test route registry
+
+Make Vitest workspace, aggregate suites, local CI selector, and routing checks consume one route registry.
+
+## Scope
+
+- In scope: Make Vitest workspace, aggregate suites, local CI selector, and routing checks consume one route registry.
+- Out of scope: unrelated refactors not required for "AP-07: Introduce unified test route registry".
+
+## Verification
+
+- State: ok
+- Note: Verified: unified test route registry drives Vitest projects, aggregate suites, local CI selector, and routing checks with 331 tests / 10 primary routes.
+- Full verification checklist lives in local review.md.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-01T08:18:15.671Z
+- Branch: task/202605010645-D8EM8Y/test-route-registry
+- Head: 4a5439e0aad7
+
+```text
+No changes detected.
+```
+
+</details>
