+++ .agentplane/tasks/202605010645-D8EM8Y/pr/github-title.txt
@@ -1,0 +1,1 @@
+AP-07: Introduce unified test route registry (D8EM8Y)
