+++ .agentplane/tasks/202605010645-D8EM8Y/pr/review.md
@@ -1,0 +1,57 @@
+# PR Review
+
+Created: 2026-05-01T08:18:15.671Z
+Branch: task/202605010645-D8EM8Y/test-route-registry
+
+## Summary
+
+AP-07: Introduce unified test route registry
+
+Make Vitest workspace, aggregate suites, local CI selector, and routing checks consume one route registry.
+
+## Scope
+
+- In scope: Make Vitest workspace, aggregate suites, local CI selector, and routing checks consume one route registry.
+- Out of scope: unrelated refactors not required for "AP-07: Introduce unified test route registry".
+
+## Verification
+
+### Plan
+
+1. Run `node scripts/check-vitest-projects.mjs`. Expected: it succeeds and confirms the requested outcome for this task.
+2. Review the changed artifact or behavior for the `code` task. Expected: the requested outcome is visible and matches the approved scope.
+3. Compare the final result against the task summary and touched scope. Expected: remaining follow-up is either resolved or explicit in ## Findings.
+
+### Current Status
+
+- State: ok
+- Note: Verified: unified test route registry drives Vitest projects, aggregate suites, local CI selector, and routing checks with 331 tests / 10 primary routes.
+
+## Risks
+
+- Risk level: not recorded
+- Breaking change: no
+
+### Rollback
+
+- Revert task-related commit(s).
+- Re-run required checks to confirm rollback safety.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-01T08:18:15.671Z
+- Branch: task/202605010645-D8EM8Y/test-route-registry
+- Head: 4a5439e0aad7
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
