+++ scripts/run-vitest-suite.mjs
@@ -1,182 +1,17 @@
 import { execFileSync, spawnSync } from "node:child_process";
-import { existsSync, readdirSync } from "node:fs";
+import { existsSync } from "node:fs";
 import path from "node:path";
-import { fileURLToPath, pathToFileURL } from "node:url";
+import { pathToFileURL } from "node:url";
 
-const REPO_ROOT = path.resolve(path.dirname(fileURLToPath(import.meta.url)), "..");
+import {
+  TEST_ROUTE_REGISTRY_ROOT as REPO_ROOT,
+  VITEST_SUITES as ROUTE_SUITES,
+} from "./lib/test-route-registry.mjs";
+export { listVitestSuiteFiles } from "./lib/test-route-registry.mjs";
+
 const VITEST_TIMEOUT_MS = "60000";
 export const VITEST_CHUNK_TIMEOUT_MS = 10 * 60 * 1000;
-
-const PRECOMMIT_FILES = [
-  "packages/agentplane/src/commands/shared/network-approval.test.ts",
-  "packages/agentplane/src/commands/shared/approval-requirements.test.ts",
-  "packages/agentplane/src/commands/task/shared.unit.test.ts",
-  "packages/agentplane/src/commands/task/shared.verify-steps.test.ts",
-  "packages/agentplane/src/commands/task/warn-owner.unit.test.ts",
-  "packages/agentplane/src/commands/shared/comment-format.test.ts",
-  "packages/agentplane/src/shared/errors.test.ts",
-  "packages/agentplane/src/cli/spec/parse.test.ts",
-  "packages/agentplane/src/cli/spec/help-render.test.ts",
-  "packages/agentplane/src/cli/spec/registry.test.ts",
-  "packages/agentplane/src/cli/spec/suggest.test.ts",
-  "packages/agentplane/src/cli/output.test.ts",
-  "packages/agentplane/src/cli/prompts.test.ts",
-  "packages/agentplane/src/cli/error-map.test.ts",
-  "packages/agentplane/src/backends/task-backend/redmine/env.test.ts",
-  "packages/core/src/config/execution-profile.test.ts",
-  "packages/core/src/config/config.test.ts",
-];
-
-const PLATFORM_CRITICAL_FILES = [
-  "packages/agentplane/src/commands/shared/pr-meta.test.ts",
-  "packages/agentplane/src/commands/scenario/impl/commands.test.ts",
-  "packages/agentplane/src/cli/run-cli.core.init.test.ts",
-  "packages/agentplane/src/cli/run-cli.core.init.branch-pr.test.ts",
-  "packages/agentplane/src/cli/run-cli.core.init.validation-conflicts.test.ts",
-  "packages/agentplane/src/cli/run-cli.core.upgrade.test.ts",
-];
-
-const BACKEND_CRITICAL_FILES = [
-  "packages/agentplane/src/backends/task-backend.test.ts",
-  "packages/agentplane/src/backends/task-backend.local.test.ts",
-  "packages/agentplane/src/backends/task-backend.redmine.cache.test.ts",
-  "packages/agentplane/src/backends/task-backend.redmine.docs.test.ts",
-  "packages/agentplane/src/backends/task-backend.redmine.mapping.test.ts",
-  "packages/agentplane/src/backends/task-backend.redmine.remote.test.ts",
-  "packages/agentplane/src/backends/task-backend.redmine.write.test.ts",
-  "packages/agentplane/src/backends/task-backend.load.test.ts",
-  "packages/agentplane/src/backends/task-backend/redmine/env.test.ts",
-  "packages/agentplane/src/commands/backend.test.ts",
-  "packages/agentplane/src/commands/shared/task-backend.test.ts",
-  "packages/agentplane/src/commands/task/export.unit.test.ts",
-  "packages/agentplane/src/commands/task/migrate-doc.test.ts",
-  "packages/agentplane/src/commands/doctor.fast.test.ts",
-  "packages/agentplane/src/cli/run-cli.core.backend-sync.test.ts",
-  "packages/agentplane/src/cli/run-cli.core.tasks.create.test.ts",
-  "packages/agentplane/src/cli/run-cli.core.tasks.incidents.test.ts",
-  "packages/agentplane/src/cli/run-cli.core.tasks.lifecycle.test.ts",
-  "packages/agentplane/src/cli/run-cli.core.tasks.update-scrub.test.ts",
-];
-
-const RELEASE_CRITICAL_FILES = [
-  "packages/agentplane/src/cli/release-critical-lifecycle.test.ts",
-  "packages/agentplane/src/cli/release-recovery-script.test.ts",
-  "packages/agentplane/src/cli/release-smoke.test.ts",
-  "packages/agentplane/src/cli/cli-smoke.test.ts",
-];
-
-const SIGNIFICANT_COVERAGE_FILES = [
-  "packages/agentplane/src/commands/guard/impl/allow.test.ts",
-  "packages/agentplane/src/commands/guard/impl/close-message.test.ts",
-  "packages/agentplane/src/commands/guard/impl/commands.unit.test.ts",
-  "packages/agentplane/src/commands/guard/impl/policy.test.ts",
-  "packages/agentplane/src/commands/guard/impl/comment-commit.test.ts",
-  "packages/agentplane/src/cli/run-cli.core.guard.test.ts",
-  "packages/agentplane/src/cli/run-cli.core.guard.commit-wrapper.test.ts",
-];
-
-function normalizeRepoPath(value) {
-  return value.split(path.sep).join("/");
-}
-
-function listFiles(relativeDir) {
-  const root = path.join(REPO_ROOT, relativeDir);
-  const files = [];
-  const pending = [root];
-  while (pending.length > 0) {
-    const current = pending.pop();
-    if (!current) continue;
-    for (const entry of readdirSync(current, { withFileTypes: true })) {
-      const absolute = path.join(current, entry.name);
-      if (entry.isDirectory()) {
-        pending.push(absolute);
-      } else if (entry.isFile()) {
-        files.push(normalizeRepoPath(path.relative(REPO_ROOT, absolute)));
-      }
-    }
-  }
-  return files.toSorted((a, b) => a.localeCompare(b));
-}
-
-function discoverTests(relativeDirs, predicate) {
-  const files = new Set();
-  for (const relativeDir of relativeDirs) {
-    for (const filePath of listFiles(relativeDir)) {
-      if (predicate(filePath)) files.add(filePath);
-    }
-  }
-  return [...files].toSorted((a, b) => a.localeCompare(b));
-}
-
-const WORKFLOW_COVERAGE_FILES = discoverTests(
-  ["packages/agentplane/src/workflow-runtime", "packages/agentplane/src/harness"],
-  (filePath) => filePath.endsWith(".test.ts"),
-);
-
-const RELEASE_CI_BASE_FILES = discoverTests(["packages"], (filePath) => {
-  if (!filePath.endsWith(".test.ts")) return false;
-  if (!filePath.includes("/src/")) return false;
-  return ![
-    /\/cli-smoke\.test\.ts$/,
-    /\/release-recovery-script\.test\.ts$/,
-    /\/run-cli\.core\.init(?:\..+)?\.test\.ts$/,
-    /\/run-cli\.core\.upgrade\.test\.ts$/,
-    /\/run-cli\.core\.backend-sync\.test\.ts$/,
-  ].some((pattern) => pattern.test(filePath));
-});
-
-export const SUITES = {
-  "backend-critical": {
-    files: BACKEND_CRITICAL_FILES,
-    maxWorkers: "4",
-    pool: "forks",
-  },
-  "platform-critical": {
-    files: PLATFORM_CRITICAL_FILES,
-    maxWorkers: "4",
-    pool: "forks",
-  },
-  precommit: {
-    files: PRECOMMIT_FILES,
-    maxWorkers: "4",
-    pool: "threads",
-  },
-  "release-ci-base": {
-    chunkSize: 10,
-    files: RELEASE_CI_BASE_FILES,
-    isolatedPatterns: [/\/run-cli\.core\.pr-flow\./],
-    maxWorkers: "4",
-    pool: "forks",
-  },
-  "release-critical": {
-    files: RELEASE_CRITICAL_FILES,
-    maxWorkers: "4",
-    pool: "forks",
-    testTimeout: "120000",
-    hookTimeout: "120000",
-  },
-  "significant-coverage": {
-    files: SIGNIFICANT_COVERAGE_FILES,
-    maxWorkers: "4",
-    pool: "threads",
-  },
-  "workflow-coverage": {
-    files: WORKFLOW_COVERAGE_FILES,
-    maxWorkers: "4",
-    pool: "threads",
-  },
-};
-
-export function listVitestSuiteFiles() {
-  return Object.fromEntries(
-    Object.entries(SUITES)
-      .map(([suiteName, suite]) => [
-        suiteName,
-        [...suite.files].toSorted((a, b) => a.localeCompare(b)),
-      ])
-      .toSorted(([a], [b]) => a.localeCompare(b)),
-  );
-}
+export const SUITES = ROUTE_SUITES;
 
 function printHelp() {
   process.stdout.write(`Usage: node scripts/run-vitest-suite.mjs <suite> [vitest args...]
